# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pllay On Edge Athlete Edition API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weekly-reflections/{weekNumber}:
    get:
      operationId: getWeeklyReflection
      tags: [weekly]
      summary: Get a specific week's reflection
      parameters:
        - name: weekNumber
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Weekly reflection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReflection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      operationId: saveWeeklyReflection
      tags: [weekly]
      summary: Save a weekly reflection (upsert)
      parameters:
        - name: weekNumber
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeeklyReflectionInput"
      responses:
        "200":
          description: Reflection saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReflection"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    WeeklyReflection:
      type: object
      required: [id, athleteId, weekNumber]
      properties:
        id:
          type: integer
        athleteId:
          type: integer
        weekNumber:
          type: integer
        completedAt:
          type: ["string", "null"]
        effort:
          type: ["integer", "null"]
        focus:
          type: ["integer", "null"]
        consistency:
          type: ["integer", "null"]
        recovery:
          type: ["integer", "null"]
        ownership:
          type: ["integer", "null"]
        bestMoment:
          type: ["string", "null"]
        biggestChallenge:
          type: ["string", "null"]
        keyLearning:
          type: ["string", "null"]
        focusAnswer:
          type: ["string", "null"]
        implWhen:
          type: ["string", "null"]
        implWhere:
          type: ["string", "null"]
        implHow:
          type: ["string", "null"]
    WeeklyReflectionInput:
      type: object
      properties:
        effort:
          type: integer
        focus:
          type: integer
        consistency:
          type: integer
        recovery:
          type: integer
        ownership:
          type: integer
        bestMoment:
          type: string
        biggestChallenge:
          type: string
        keyLearning:
          type: string
        focusAnswer:
          type: string
        implWhen:
          type: string
        implWhere:
          type: string
        implHow:
          type: string
        completed:
          type: boolean
...