# lib/api-zod/src/generated/types/weeklyReflection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pllay On Edge Athlete Edition API
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyReflection {
  id: number;
  athleteId: number;
  weekNumber: number;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  effort?: number | null;
  /** @nullable */
  focus?: number | null;
  /** @nullable */
  consistency?: number | null;
  /** @nullable */
  recovery?: number | null;
  /** @nullable */
  ownership?: number | null;
  /** @nullable */
  bestMoment?: string | null;
  /** @nullable */
  biggestChallenge?: string | null;
  /** @nullable */
  keyLearning?: string | null;
  /** @nullable */
  focusAnswer?: string | null;
  /** @nullable */
  implWhen?: string | null;
  /** @nullable */
  implWhere?: string | null;
  /** @nullable */
  implHow?: string | null;
}